mv annotation ctx to private
